refact(firebase): bundling firebase functions
