+++ examples/v0.77.0/hot-updater.config.ts
@@ -1,5 +1,6 @@
+
 import { metro } from "@hot-updater/metro";
-import { supabaseDatabase, supabaseStorage } from "@hot-updater/supabase";
+import { firebaseStorage, firebaseDatabase } from "@hot-updater/firebase";
 import { defineConfig } from "hot-updater";
 import "dotenv/config";
 
@@ -7,13 +8,13 @@ export default defineConfig({
   build: metro({
     enableHermes: true,
   }),
-  storage: supabaseStorage({
-    supabaseUrl: process.env.HOT_UPDATER_SUPABASE_URL!,
-    supabaseAnonKey: process.env.HOT_UPDATER_SUPABASE_ANON_KEY!,
-    bucketName: process.env.HOT_UPDATER_SUPABASE_BUCKET_NAME!,
+  storage: firebaseStorage({
+    apiKey: process.env.HOT_UPDATER_FIREBASE_API_KEY,
+    projectId: process.env.HOT_UPDATER_FIREBASE_PROJECT_ID,
+    storageBucket: process.env.HOT_UPDATER_FIREBASE_STORAGE_BUCKET,
   }),
-  database: supabaseDatabase({
-    supabaseUrl: process.env.HOT_UPDATER_SUPABASE_URL!,
-    supabaseAnonKey: process.env.HOT_UPDATER_SUPABASE_ANON_KEY!,
+  database: firebaseDatabase({
+    apiKey: process.env.HOT_UPDATER_FIREBASE_API_KEY,
+    projectId: process.env.HOT_UPDATER_FIREBASE_PROJECT_ID,
   }),
 });
