+++ examples/v0.77.0/package.json
@@ -18,10 +18,12 @@
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
     "@babel/runtime": "^7.25.0",
-    "@hot-updater/aws": "0.12.7",
-    "@hot-updater/cloudflare": "0.12.7",
-    "@hot-updater/metro": "0.12.7",
-    "@hot-updater/supabase": "0.12.7",
+    "@google-cloud/functions-framework": "^3.4.6",
+    "@hot-updater/aws": "0.12.4",
+    "@hot-updater/cloudflare": "0.12.4",
+    "@hot-updater/firebase": "workspace:*",
+    "@hot-updater/metro": "0.12.4",
+    "@hot-updater/supabase": "0.12.4",
     "@react-native-community/cli": "15.0.1",
     "@react-native-community/cli-platform-android": "15.0.1",
     "@react-native-community/cli-platform-ios": "15.0.1",
@@ -38,7 +40,8 @@
     "@types/react-test-renderer": "^18.0.0",
     "dotenv": "^16.4.5",
     "eslint": "^8.19.0",
-    "hot-updater": "0.12.7",
+    "firebase-tools": "^13.32.0",
+    "hot-updater": "0.12.4",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
     "react-native-dotenv": "^3.4.11",
