+++ plugins/firebase/firebase/firebase.json
@@ -13,8 +13,7 @@
         "firebase-debug.log",
         "firebase-debug.*.log",
         "*.local"
-      ],
-      "predeploy": ["npm --prefix \"$RESOURCE_DIR\" run build"]
+      ]
     }
   ],
   "storage": {
