+++ plugins/firebase/firebase/functions/_package.json
@@ -1,33 +1,8 @@
 {
   "name": "functions",
-  "type": "module",
-  "scripts": {
-    "build": "tsc",
-    "build:watch": "tsc --watch",
-    "serve": "npm run build && firebase emulators:start --only functions",
-    "shell": "npm run build && firebase functions:shell",
-    "start": "npm run shell",
-    "deploy": "firebase deploy --only functions",
-    "logs": "firebase functions:log",
-    "test": "vitest"
-  },
-  "engines": {
-    "node": "22"
-  },
-  "main": "lib/index.js",
+  "engines": { "node": "22" },
   "dependencies": {
-    "@hot-updater/core": "^0.12.3",
-    "@hot-updater/js": "^0.12.3",
-    "firebase-admin": "^12.6.0",
-    "firebase-functions": "^6.0.1"
+    "firebase-functions": "^6.3.2"
   },
-  "devDependencies": {
-    "@types/node": "^22.13.5",
-    "@types/semver": "^7.5.8",
-    "firebase-functions-test": "^3.1.0",
-    "semver": "^7.6.3",
-    "typescript": "^5.7.3",
-    "vitest": "^3.0.7"
-  },
-  "private": true
+  "main": "index.cjs"
 }
