+++ plugins/firebase/firebase/functions/tsconfig.json
@@ -12,5 +12,5 @@
     "lib": ["es2020", "dom"]
   },
   "compileOnSave": true,
-  "include": ["src"]
+  "include": ["src", "index.ts"]
 }
